Continue PNG work; tweak fetch code

+++ src/main/java/com/github/webicitybrowser/codec/png/PNGResult.java
@@ -1,5 +1,5 @@
 package com.github.webicitybrowser.codec.png;
 
-public record PNGResult(int width, int height, int[] pixels) {
+public record PNGResult(int width, int height, byte[] pixels) {
 	
 }
